% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\docType{methods}
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Find a solution to the knapsack problem with the greedy packing algorithm}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame containing rows for items to be packed with columns 'w' for weight and 'v' for value}

\item{W}{The maximum weight of the knapsack}
}
\value{
A list containing the maximum value of the knapsack and a vector of elements contained in the knapsack
}
\description{
Find a solution to the knapsack problem with the greedy packing algorithm
}
