\name{solveKnapsack-package}
\alias{solveKnapsack-package}
\alias{solveKnapsack}
\docType{package}
\title{
\packageTitle{solveKnapsack}
}
\description{
\packageDescription{solveKnapsack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{solveKnapsack}
\packageIndices{solveKnapsack}
~~ An overview of how to use the ~~
~~ package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{solveKnapsack}

Maintainer: \packageMaintainer{solveKnapsack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, ~~
~~ one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
